requirements.txt
appdirs==1.4.4
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
black==19.10b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
flake8==4.0.1
Flask==2.0.2
flask-ngrok==0.0.25
fonttools==4.28.5
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
ngrok==0.0.1
numpy==1.22.0
packaging==21.3
pathspec==0.7.0
Pillow==9.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
regex @ file:///C:/ci/regex_1629302316714/work
requests==2.27.1
six==1.16.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.10.1
torchvision==0.11.2
typed-ast @ file:///C:/ci/typed-ast_1624953785070/work
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wincertstore==0.2
